Add casts to strlen where appropriate

+++ src/nsh_cmd.c
@@ -46,7 +46,7 @@ nsh_cmd_t* nsh_cmd_find_matching(nsh_cmd_array_t* cmds, const char* partial_name
 
 nsh_cmd_t* nsh_cmd_find(nsh_cmd_array_t* cmds, const char* name)
 {
-    return nsh_cmd_find_matching(cmds, name, strlen(name));
+    return nsh_cmd_find_matching(cmds, name, (int)strlen(name));
 }
 
 int nsh_cmd_register(nsh_cmd_array_t* cmds, const char* name, nsh_cmd_handler_t handler)
